+++ docs/source/retry-policy/default.md
@@ -13,7 +13,7 @@ To use in `Session`:
 # async fn check_only_compiles() -> Result<(), Box<dyn Error>> {
 use scylla::{Session, SessionBuilder};
 use scylla::execution::ExecutionProfile;
-use scylla::transport::retry_policy::DefaultRetryPolicy;
+use scylla::execution::retry_policy::DefaultRetryPolicy;
 
 let handle = ExecutionProfile::builder()
     .retry_policy(Arc::new(DefaultRetryPolicy::new()))
@@ -38,7 +38,7 @@ To use in a [simple query](../queries/simple.md):
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
 use scylla::query::Query;
 use scylla::execution::ExecutionProfile;
-use scylla::transport::retry_policy::DefaultRetryPolicy;
+use scylla::execution::retry_policy::DefaultRetryPolicy;
 
 // Create a Query manually and set the retry policy
 let mut my_query: Query = Query::new("INSERT INTO ks.tab (a) VALUES(?)");
@@ -67,7 +67,7 @@ To use in a [prepared query](../queries/prepared.md):
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
 use scylla::prepared_statement::PreparedStatement;
 use scylla::execution::ExecutionProfile;
-use scylla::transport::retry_policy::DefaultRetryPolicy;
+use scylla::execution::retry_policy::DefaultRetryPolicy;
 
 // Create PreparedStatement manually and set the retry policy
 let mut prepared: PreparedStatement = session
