+++ scylla/src/execution/execution_profile.rs
@@ -216,7 +216,7 @@ pub(crate) mod defaults {
 /// # Example
 ///
 /// ```
-/// # use scylla::transport::retry_policy::FallthroughRetryPolicy;
+/// # use scylla::execution::retry_policy::FallthroughRetryPolicy;
 /// # use scylla::execution::ExecutionProfile;
 /// # use scylla::statement::Consistency;
 /// # use std::sync::Arc;
@@ -297,12 +297,12 @@ impl ExecutionProfileBuilder {
     }
 
     /// Sets the [`RetryPolicy`] to use by default on queries.
-    /// The default is [DefaultRetryPolicy](crate::transport::retry_policy::DefaultRetryPolicy).
+    /// The default is [DefaultRetryPolicy](crate::execution::retry_policy::DefaultRetryPolicy).
     /// It is possible to implement a custom retry policy by implementing the trait [`RetryPolicy`].
     ///
     /// # Example
     /// ```
-    /// use scylla::transport::retry_policy::DefaultRetryPolicy;
+    /// use scylla::execution::retry_policy::DefaultRetryPolicy;
     /// # use scylla::execution::ExecutionProfile;
     /// # use std::sync::Arc;
     /// # fn example() -> Result<(), Box<dyn std::error::Error>> {
@@ -353,7 +353,7 @@ impl ExecutionProfileBuilder {
     ///
     /// # Example
     /// ```
-    /// use scylla::transport::retry_policy::DefaultRetryPolicy;
+    /// use scylla::execution::retry_policy::DefaultRetryPolicy;
     /// # use scylla::execution::ExecutionProfile;
     /// # use std::sync::Arc;
     /// # fn example() -> Result<(), Box<dyn std::error::Error>> {
