+++ scylla/src/lib.rs
@@ -293,7 +293,7 @@ pub use session::CloudSessionBuilder;
 pub use cluster::host_filter;
 pub use execution::execution_profile;
 pub use execution::load_balancing;
+pub use execution::retry_policy;
 pub use execution::speculative_execution;
-pub use transport::retry_policy;
 
 pub use transport::metrics::{Metrics, MetricsError};
