+++ scylla/src/transport/iterator.rs
@@ -34,6 +34,7 @@ use crate::cluster::ClusterData;
 use crate::cluster::NodeRef;
 use crate::connection::{Connection, NonErrorQueryResponse, QueryResponse};
 use crate::execution::load_balancing::{self, RoutingInfo};
+use crate::execution::retry_policy::{QueryInfo, RetryDecision, RetrySession};
 use crate::frame::response::{
     result,
     result::{ColumnSpec, Row},
@@ -42,7 +43,6 @@ use crate::history::{self, HistoryListener};
 use crate::statement::{prepared_statement::PreparedStatement, query::Query};
 use crate::transport::errors::ProtocolError;
 use crate::transport::metrics::Metrics;
-use crate::transport::retry_policy::{QueryInfo, RetryDecision, RetrySession};
 use tracing::{trace, trace_span, warn, Instrument};
 use uuid::Uuid;
