+++ scylla/src/transport/mod.rs
@@ -1,4 +1,3 @@
-pub mod downgrading_consistency_retry_policy;
 pub mod errors;
 pub mod iterator;
 pub mod legacy_query_result;
@@ -6,7 +5,6 @@ pub mod locator;
 pub(crate) mod metrics;
 pub mod partitioner;
 pub mod query_result;
-pub mod retry_policy;
 
 pub use crate::frame::Authenticator;
 pub use scylla_cql::frame::request::query::{PagingState, PagingStateResponse};
